feat(kre-py tests): Refactor fixtures to a common file
